Add jupyter nb results Closes #3 Closes #1

+++ Models/NaiveBayes.py
@@ -7,18 +7,23 @@ import os, pickle
 
 
 def run(data):
+    if not os.path.isfile('Data/pickles/mb_grid'):
 
-    params = {
-        'alpha': [0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09,
-                  0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 5, 10]
-    }
+        params = {
+            'alpha': [0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09,
+                      0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 5, 10]
+        }
 
-    grid = GridSearchCV(MultinomialNB(), param_grid=params, cv=5, verbose=3)
+        grid = GridSearchCV(MultinomialNB(), param_grid=params, cv=5, verbose=3)
 
-    grid.fit(data['x_train'], data['y_train'])
+        grid.fit(data['x_train'], data['y_train'])
 
-    with open('Data/pickles/mb_grid', 'wb') as file:
-        pickle.dump(grid, file)
+        with open('Data/pickles/mb_grid', 'wb') as file:
+            pickle.dump(grid, file)
+
+    else:
+        with open('Data/pickles/mb_grid', 'rb') as file:
+            grid = pickle.load(file)
 
     model = grid.best_estimator_
 
@@ -26,4 +31,4 @@ def run(data):
 
     print(accuracy_score(data['y_test'], y_pred))
 
-    return confusion_matrix(data['y_test'], y_pred)
+    return {'acc': confusion_matrix(data['y_test'], y_pred), 'grid': grid}
